Refactor LLM methods into library file
